+++ Doctrine/Phpcr/AutoRouteListener.php
@@ -10,12 +10,12 @@
  */
 
 
-namespace Symfony\Cmf\Bundle\RoutingAutoBundle\EventListener;
+namespace Symfony\Cmf\Bundle\RoutingAutoBundle\Doctrine\Phpcr;
 
 use Doctrine\Common\Persistence\Event\ManagerEventArgs;
 use Doctrine\ODM\PHPCR\DocumentManager;
 use Symfony\Component\DependencyInjection\ContainerInterface;
-use Symfony\Cmf\Bundle\RoutingAutoBundle\Document\AutoRoute;
+use Symfony\Cmf\Bundle\RoutingAutoBundle\Model\AutoRoute;
 
 /**
  * Doctrine PHPCR ODM listener for maintaining automatic routes.